// src/app/interfaces/justificacion.ts
export interface Ijustifica {
    id:String
    rut:String
    asignatura: String
    description: String
    fecha: Date
    profesor: String
}
export interface newjustifica {
    id:String
    rut:String
    asignatura: String
    description: String
    fecha: Date
    profesor: String
}